Add files via upload
Title: Create a histogram or bar chart to visualize the distribution of data in a dataset

Description: This repository contains a Python script for creating a histogram or bar chart to visualize the distribution of data in a dataset. The script can be used to create a histogram for any numeric column in a dataset, or a bar chart for any categorical column in a dataset.

Usage: To use the script, simply install the required Python packages and then run the script with the path to the dataset and the name of the column to visualize as the arguments. The script will generate a plot file in the same directory as the dataset.