Add files via upload
Title: Use pivot tables to summarize data in a dataset

Description: This repository contains a Python script for using pivot tables to summarize data in a dataset. The script can be used to create pivot tables for any dataset, regardless of its size or complexity.

Usage: To use the script, simply install the required Python packages and then run the script with the path to the dataset as the argument. The script will generate a pivot table file in the same directory as the dataset.